Support specifying mode on individual screen
Allow a screen to define it's `mode` and that will be used in place of the
navigator's default.

Requires navigator to have a headerMode of `screen` so that the headers behave
properly.

Still has the problem of the screen you are transitioning to slightly going left
and I'm not sure where that happens.